Should use a dummy file to prevent FF merges.
I may maintain a private Scrapy project. And to prevent merging private code
with public code, I will let this commit slip into the tree. My private branch
will not have it and, therefore, I will not be able to merge it with public
code with a FF merge. This is a protection against my own sleep deprivation.