Should use autocommit with SQLite 3.
This way, the connection commits after every executed statement. This
seems to be the safest way to handle transactions.

Also logging at debug level the time it takes to ask for the first
user input so we can better evaluate user experience.

Lastly, also creating an index on playing_data (score_delta) because
this simplifies the query plan for summing all score_delta values.